more implementation and adjustment.

+++ src/OnlineAssessment.Domain/Examination/PaperQuestion.cs
@@ -13,13 +13,13 @@ namespace OnlineAssessment.Domain
         public Guid PaperQuestionId { get; set; }
         public int QuestionIndex { get; set; }
         public int Score { get; set; }
-        public float QuestionDegree { get; set; }
+        public double QuestionDegree { get; set; }
         public QuestionForm QuestionForm { get; set; }
         public string QuestionBody { get; set; }
         public string ReferenceRightAnswer { get; set; }
         public ICollection<PaperQuestionOption> QuestionOptions { get; set; }
 
-        public float AvarageDegree {
+        public double AvarageDegree {
             get { return QuestionDegree*Score; }
         }
     }
